java.lang.IllegalArgumentException: voidcannot be converted to an Element

+++ app/src/main/java/net/mnsam/antnote/main/database/dao/NoteDao.kt
@@ -5,7 +5,7 @@ import android.arch.persistence.room.Dao
 import android.arch.persistence.room.Insert
 import android.arch.persistence.room.OnConflictStrategy
 import android.arch.persistence.room.Query
-import net.mnsam.antnote.main.database.model.Note
+import net.mnsam.antnote.main.database.entity.Note
 
 /**
  * Created by Mochamad Noor Syamsu on 12/27/17.
@@ -14,7 +14,7 @@ import net.mnsam.antnote.main.database.model.Note
 @Dao
 interface NoteDao {
     @Insert(onConflict = OnConflictStrategy.FAIL)
-    fun insert(note: Note): Int
+    fun insert(note: Note): Long
 
     @Query("SELECT * from notes ORDER BY id ASC")
     fun getAllNotes(): LiveData<MutableList<Note>>
